Add metalink support (#3683)
This adds support for CVMFS_METALINK_URL, CVMFS_EXTERNAL_METALINK, and CVMFS_METALINK_RESET_AFTER. Much of the code supporting hosts is reused for metalinks. The METALINK variables are much like their SERVER counterparts, except they do not geosort, they take precedent over the SERVER settings, and if there are rfc6249 Link: header(s) they are used to set the list of servers. If a Link header is received, then the metalink server is not used until CVMFS_METALINK_AFTER_DELAY seconds later, which defaults to the value of CVMFS_PROXY_RESET_AFTER.

I don't yet see a use case for CVMFS_METALINK_URL but it was trivial to add in addition to CVMFS_EXTERNAL_METALINK so I thought it might as well be included.


* fix bogus changelog entry

* add metalink support infrastructure, based on host support

* request CVMFS_EXTERNAL_METALINK hosts (if set) before CVMFS_EXTERNAL_URL hosts

* process the LINK: header

* add metalink test

* cloudtest fixes

* fix uninitialized metalink timestamp, add helper method for code clarity

* test server won't run with python2 (affects centos7)

* clang-tidy fixes

* reduce casting by making "current" an int instead of unsigned

* clang-tidy

---------

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>