[ci] fix containerized cloudtests for cgroup v2
Ubuntu 22.04 and newer use cgroup v2, which changes user namespace
remapping. The test container now exits immediately after /sbin/init,
instead of running in the background [1, 2]. One fix is the docker
setting `"userns-remap": "default"`, but that doesn't work with
`--privileged`.  So I use `--cgroupns host` instead, which shouldn't
be an issue on transient test vms.

The dockerized test setup really should avoid systemd altogether, but
we need to run httpd as a service container and modify a few checks for
systemd.

[1] https://github.com/geerlingguy/docker-fedora37-ansible/issues/1
[2] https://github.com/docker/for-mac/issues/6288