[gw] Early recursion return when adding nested catalogs (#3565)
Fixes #3562: cvmfs_receiver: grafting inefficiency when adding new directories with nested catalogs

"""
This problem manifests when the gateway lease path does not already exist in the repo - new nested catalogs are unnecessarily recursed into during the merge.

consider a directory hierarchy /TOPDIR/1/2/3/4/4/5/6/7/8/9/10 being created on an existing /TOPDIR with catalog. A catalog is added on each directory. The change is published via a gateway, with lease TOPDIR/1. The cvmfs_receiver commit processor will unexpectedly recurse down through all of the catalogs in the merge, down to ..../10.
In a subsequent commit that touches all of these directories, where the lease path already exists, the commit processor will only wrangle the catalogs for /, /TOPDIR, /TOPDIR/1, as expected.

This initial, apparently unnecessary, recursion marked slows down the critical path commit of new directories into a repo

Looks like this might be a case not handled by the previous nested catalog optimisation #2707
"""

---------

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>