Remove vendored golang toolchain and bump version requirement to 1.21 (#3818)
The vendored version is becoming outdated, and it seems easy and unproblematic to use the golang toolchain from the system or download it. I've added a few lines that download and install 1.21 in the build images for platforms with older go versions (  Debian <= 12 and Ubuntu < 24.04 have go < 1.21).

For centos7 I'd have to remove the BuildRequires in the rpm (Debian-based platforms don't even have a build dependency on it, just dynamically switch it on/off - to be improved), so I'd opted to just drop the platform for building the go code.