ducc: start testing and mocking cvmfs bindings (#3829)
This was a bit of a rabbithole, but ultimately very useful: In order to
be able to unittest the unpacker, I wanted to be able to unpack to a
normal directory (outside of cvmfs). Since the unpacker calls out to
cvmfs_server, this can be achieved by mocking cvmfs_server to do nothing
in a normal transaction, and a copy in a template transaction. The
sneaky chains made this a bit more complicated, especially since I want
to change that way of doing things.

There is a bit of specific setup for the tests to work, which can be
found in cvmfs_test.go. As the unpacker gets cleaned up, this will no
longer be needed, but adds some test coverage for the rewrite.