Avoid use of fusermount3 by premounting & remove fuse3 dependency (#3796)
When cvmfs2 is running as root with a uid= option, it unnecessarily invokes fusermount3. This can be avoided by doing the mount before changing the effective uid and then invoking the fuse library with a "/dev/fd/N" string in place of the mountpoint, where N is the number of a "/dev/fuse" file descriptor that was passed to a mount of fuse at the mountpoint.

This should also remove the need to depend on the fuse3 package. It requires fuse3-libs >= 3.3.0, which is available in all currently supported common operating systems. It will also remove the need for a fusermount3 apparmor profile including /cvmfs for Ubuntu 24.10+ as done in #3791.


- Fixes #3795

---------

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>