Containerized tests multistage dockerfile improvements (#3827)
Common container tests now use a multistage dockerfile. It copies the
repository from the host and uses this for the build stage. The
resulting rpm files are copied to a second stage, where they are
installed. This multistage setup ensures the large amount of build files
do not end up in the resulting container. Because the docker `COPY`
command only works inside of the build context, the build context is
changed to the repository root. The concurrent publishing tests require
the same container setup, which is why this now also use this common
dockerfile (instead of its own).

---------

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>