[test] 084-premount: fix cleanup (#3834)
The cleanup of this test was only partly implemented, and missed a $ for
CVMFS_TEST084_FUSEPREMOUNT. I don't see the necessity of moving the
premount helper binary to `/`, so I leave it in the workspace now.

To reduce the noisiness of this test, the output of some commands are
redirected to logfiles. New expected output:
```
*** creating fuse-premount
   -rwsr-xr-x. 1 root root 19K May 16 20:49 /home/sftnight/cvmfs/test/src/084-premounted/fuse_premount
*** mounting on /home/sftnight/cvmfs/test/src/084-premounted/mountpoint
*** listing repository mounted at /home/sftnight/cvmfs/test/src/084-premounted/mountpoint
*** checking dirtab in  /home/sftnight/cvmfs/test/src/084-premounted/mountpoint
*** finished successfully!
*** cleanup 
  *** unmounted /home/sftnight/cvmfs/test/src/084-premounted/mountpoint
  *** removed /home/sftnight/cvmfs/test/src/084-premounted/cache
```