[test] Fix permissions in 81 and timeouts on find (#3835)
In a new cloudtesting setup that avoids the
test/cloudtesting/platforms/*.sh scripts, there were several timeouts
that seemed to come from `find` traversing more of the repository than
intended.

The command find $repo | head -n 1 is indeed not guaranteed to stop the
find after one file, and depends on line buffering of the shell.

This uses subdirectories of the repositories to find test files, to
limit find and fix the tests. The downside is that those subdirectories
could be deleted and the test logic needs to be updated, but that should
be rare and not a problem.