[infra] Update googletest dep, use cmake FetchExternalContent to build (#3874)
Fixes the issue mentioned in https://github.com/cvmfs/cvmfs/pull/3852

With this change, it is possible to use either the system googletest
package if available or build it from source using CMake's
FetchExternalContent.

I encountered a weird CI error on MacOS in this PR; with this seemingly
unrelated change MacOS would pick up sqlite3 headers from the system. I
couldn't reproduce this on our mac build machine so I just add a
workaround to the CI.