Update release package with new signing key for Alma10 (#3890)
Like Debian-based platforms started doing a while ago, Alma 10 doesn't
accept 1024 bit signing keys anymore. For Debian we just signed the
packages twice with the old and a new key, but that does not seem to be
possible with rpms.

So I update the release package to use the same new signing key already
used for debian exclusively from Almalinux 10 on.