Add  manpages for some cvmfs commands (#3833)
Adds support for:
* `man cvmfs2`
* `man cvmfs_config`
* `man cvmfs_server`
* `man cvmfs_swissknife`

This is mainly for compliance with packaging guidelines of linux distro
repositories. The man pages don't add much more information than what is
currently available in the `--help` output, and use `help2man` in order
to keep the information in sync with the codebase.

The `--help` output needs to be improved in a future patch - in
particular cvmfs_server and swissknife should show the usage for
individual subcommands and not just one long list with all of them.