Add unittest infrastructure for FUSE callbacks (#3918)
Was fairly painful to mock everything, but at least gave some ideas for
future renovations. This approach certainly tests some implementation
details, but for the callbacks I think this is justified. The code
changes little enough that updating the test logic when it does is
reasonable.

Currently only some basic tests for `cvmfs_release` - regression tests
for the page cache tracker issues are coming next.

Fixes #3911