Fix race when using page cache tracker for chunked files (#3685)
When opening chunked files, the use of the page cache tracker is not
protected by a remount fence. As a result, the following condition can
happen:

```
Thread 1                                                                              Thread 2                                          Extern

Opens /foo, hangs just before writing to PCT
                                                                                                                                        /foo replaced by new version
                                                                                      Lookup /foo
                                                                                      gets live inode because not found in PCT
Commits entry to PCT
                                                                                      Open /foo --> PCT sets direct I/O
                                                                                      Error (EIO or EMFILE)
                                                                                      Crash in PCT Close
```

I _think_ the condition where the PCT sets the open directive to direct
I/O should not appear anymore with the race fixed. But I haven't
convinced myself that this is really the case.

Fixes #3520 
Relies on #3507
Replaces #3609