Allocate cvmfs_server lock file descriptors dynamically (#3913)
Instead of always using fd 9 for locks, open the next available file
descriptor and keep track of it.

Also after obtaining the lock check if the file has been removed by a release of the former lock holder and if so, close the lock and obtain it again.  See longer rationale in [my issue comment](https://github.com/cvmfs/cvmfs/issues/3909#issuecomment-3063304960).

This could even happen if the lock aborts on a conflict, so rather than duplicating that logic in two functions I decided to combine them into one `acquire_lock()` function.  That case could happen if process 2 opens the file, then process 1 removes and closes the file before process 2 calls `flock -n`.  It's a much less likely race condition than if process 2 waits for the lock to be freed -- in that case the problem was basically inevitable -- but it is possible.

- Fixes #3909