`cvmfs_talk metrics prometheus`  (#3944)
This adds a talk command that outputs a lot of information in a format
that can be ingested by Prometheus. The motivation is to speed up the
script in https://github.com/cvmfs-contrib/prometheus-cvmfs , which was
very slow due to the overhead of calling cvmfs_talk over and over for
each metric. Having all metrics returned by one command reduces the
runtime of the script from something like 13s to 2.5s second for 20
mounted repos.

A possible future optimization could be to offload some work, like
constructing the string, from talk.cc to cvmfs_talk.cc, and only send
the raw values in a suitable format over the socket.