cmake: better externals dir naming, re-use legacy externals folders (#3955)
>It's probably not wrong to add some logic to keep using legacy
`externals_install` directories if they exist. Otherwise a new set of
externals will be built in the new directory, but the build directory
will keep using the old one, that can be a bit confusing. Will still add
a patch

_Originally posted by @vvolkl in
https://github.com/cvmfs/cvmfs/issues/3951#issuecomment-3144853285_

This is the patch. Also uses PLATFORM_ID instead of the unnecessarily
specific DISTRO_ID where available, so the name is
`externals_install.x86_64.el9/` instead of
`externals_install.x86_64.almalinux9.6/`