Update geo api for either mmdb name (#3982)
Unfortunately I didn't fully test the new geo db download in #3967, and
we apparently didn't notice that the cloudtest 614-geoservice would have
failed. This fixes that situation, including removing the alternate
database name when switching between the two different types.

Test 614-geoservice takes about 3 minutes so it isn't exactly "quick"
but I wonder if we should include it in the quick tests anyway so it
will always run. Alternatively, maybe the Jenkins cloudtests
configuration can be changed to use the new geo database.