Setup global environment params inside the OptionsManager (#3999)
This is an attempt to solve https://github.com/cvmfs/cvmfs/issues/3971,
also a cleaner version of closed
[PR-3994](https://github.com/cvmfs/cvmfs/pull/3994)

What is happening here basically is we're transferring the
SetupGlobalEnvironmentParams() inside the OptionsManager.

The PR is marked as draft since probably some further cleanup is needed
(Now the same functionality is provided from theOptionsManager and the
Filesystem).

To test that the environment is setup properly, we provide a temporary
/etc/cvmfs/default.local that reports if CVMFS_VERSION and CVMFS_ARCH
are properly set.

---------

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>