loader: do not attempt umount when fuse exits with ret 0 (#4014)
Another possible fix for #3993 . Here we do not even attempt the umount
when the fuse loop exits with return value 0, very effectively avoiding
the complications of a possible race with another following mount
attempt.

However with this patch, a cvmfs repository needs to be manually
unmounted when the filesystem connection in /sys/fs/fuse is aborted.
Since this is rare enough and will only happen on admin intervention, it
should be fine.