[client] Skip open files on cache cleanup when possible (#4029)
On the Algorithmic side open files are handled as if they were pinned.

The tricky part is how the PosixQuotaManager (PQM) will get to know
which files are open. In this approach, open hashes are communicated to
the PQM using a dedicated magic xattr.

Firstly the PQM needs to know its managed mount points. This is
happening early, when the PQM is acquired.
To get the open hashes of a mount point, this approach introduces the
magic xattr `user.list_open_hashes`.
To get the mount points managed from the shared PQM, this approach
introduces the magic xattr `user.list_managed_mountopints`.
To get the open hashes of all the mount points associated to the shared
PQM, this approach introduces the magic xattr
`user.list_groups_open_hashes`. This last magic xattr uses exactly the
same mechanism that is employed in `DoCleanup()`.


By default on this stage the feature is turned off. To enable it add
`CVMFS_CACHE_CLEANUP_NONOPENLRU=yes` in the configuration.

---------

Signed-off-by: Georgios Christodoulis <christge@cern.ch>