Add support for xattrs with values up to 64k (#3622)
Adds support in a backward and somewhat forward compatible way. If only
small attributes are attached to a file, serialization will use the
current format understood by old readers. If one of the attributes is a
large one, the list of attributes is serialized in a new format that old
clients don't understand. Old clients won't crash but they will show
none of the attributes.

Fixes #3621

---------

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>